ML pipelines
The code in Data608Project.ipynb was rewritten and restructured as PySpark pipelines in the hope the code would run more efficiently and not be affected by the system resource errors that have prevented the programs from completing.

The dataset was formatted two ways for the pipeline. The original unbalanced dataset was used. For the balanced dataset, each category of the data was randomly sampled to create a dataset where each category had an equal number of data points.